pyproject.toml
[tool.poetry]
name = "catequese"
version = "0.1.0"
description = ""
authors = ["Thiago <thiago.costa@mateusmais.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
django-localflavor = "^3.1"
Pillow = "^9.0.0"
django-bootstrap-v5 = "^1.0.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
